requirements.txt
flask
flask_cors
google.generativeai
python-dotenv
firebase-admin